+++ src/patkit/gui/annotator_window.py
@@ -156,28 +156,35 @@ class UiMainWindow(object):
 
         main_window.setCentralWidget(self.centralwidget)
 
-        ### Menus and menu items
+        ### Menu bar
         self.menubar = QtWidgets.QMenuBar(main_window)
         self.menubar.setGeometry(QtCore.QRect(0, 0, 1087, 22))
         self.menubar.setObjectName("menubar")
 
+        ### Menus
         self.menu_file = QtWidgets.QMenu(self.menubar)
         self.menu_file.setObjectName("menu_file")
+        self.menu_export = QtWidgets.QMenu(self.menubar)
+        self.menu_export.setObjectName("menu_export")
+        self.menu_exercise = QtWidgets.QMenu(self.menubar)
+        self.menu_exercise.setObjectName("menu_exercise")
         self.menu_image = QtWidgets.QMenu(self.menubar)
-
         self.menu_image.setObjectName("menu_image")
+        self.menu_navigation = QtWidgets.QMenu(self.menubar)
+        self.menu_navigation.setObjectName("menu_navigation")
         self.menu_script = QtWidgets.QMenu(self.menubar)
         self.menu_script.setEnabled(False)
         self.menu_script.setObjectName("menu_script")
-        self.menu_navigation = QtWidgets.QMenu(self.menubar)
-        self.menu_navigation.setObjectName("menu_navigation")
-        self.menu_export = QtWidgets.QMenu(self.menubar)
-        self.menu_export.setObjectName("menu_export")
         main_window.setMenuBar(self.menubar)
+
+        ### Statusbar
         self.statusbar = QtWidgets.QStatusBar(main_window)
         self.statusbar.setObjectName("statusbar")
         main_window.setStatusBar(self.statusbar)
 
+        ## Menu items
+
+        ### File menu
         self.actionNew = QtGui.QAction(main_window)
         self.actionNew.setObjectName("actionNew")
         self.action_open = QtGui.QAction(main_window)
@@ -195,20 +202,46 @@ class UiMainWindow(object):
         self.action_quit = QtGui.QAction(main_window)
         self.action_quit.setObjectName("action_quit")
 
-        self.actionShow_interpreter = QtGui.QAction(main_window)
-        self.actionShow_interpreter.setObjectName("actionShow_interpreter")
-        self.actionRun_file = QtGui.QAction(main_window)
-        self.actionRun_file.setObjectName("actionRun_file")
+        self.menu_file.addAction(self.action_open)
+        self.menu_file.addAction(self.action_save_current_textgrid)
+        self.menu_file.addAction(self.action_save_all_textgrids)
+        self.menu_file.addAction(self.action_save_all)
+        self.menu_file.addSeparator()
+        self.menu_file.addAction(self.action_quit)
 
-        self.actionNext = QtGui.QAction(main_window)
-        self.actionNext.setObjectName("actionNext")
-        self.actionPrevious = QtGui.QAction(main_window)
-        self.actionPrevious.setObjectName("actionPrevious")
-        self.actionNext_Frame = QtGui.QAction(main_window)
-        self.actionNext_Frame.setObjectName("actionNext_Frame")
-        self.actionPrevious_Frame = QtGui.QAction(main_window)
-        self.actionPrevious_Frame.setObjectName("actionPrevious_Frame")
+        ### Exercise menu actions
+        self.action_run_as_exercise = QtGui.QAction(main_window)
+        self.action_run_as_exercise.setObjectName("action_run_as_exercise")
+        self.action_create_exercise = QtGui.QAction(main_window)
+        self.action_create_exercise.setObjectName("action_create_exercise")
+        self.action_open_exercise = QtGui.QAction(main_window)
+        self.action_open_exercise.setObjectName("action_open_exercise")
+        self.action_open_answer = QtGui.QAction(main_window)
+        self.action_open_answer.setObjectName("action_open_answer")
+        self.action_save_answer = QtGui.QAction(main_window)
+        self.action_save_answer.setObjectName("action_save_answer")
+        self.action_compare_to_example = QtGui.QAction(main_window)
+        self.action_compare_to_example.setObjectName("action_compare_to_example")
+        self.action_show_example = QtGui.QAction(main_window)
+        self.action_show_example.setObjectName("action_show_example")
 
+        self.menu_exercise.addAction(self.action_run_as_exercise)
+        self.menu_exercise.addAction(self.action_create_exercise)
+        self.menu_exercise.addAction(self.action_open_exercise)
+        self.menu_exercise.addSeparator()
+        self.menu_exercise.addAction(self.action_open_answer)
+        self.menu_exercise.addAction(self.action_save_answer)
+        self.menu_exercise.addSeparator()
+        self.menu_exercise.addAction(self.action_compare_to_example)
+        self.menu_exercise.addAction(self.action_show_example)
+
+        # TODO: 0.18.1: Implement this?
+        self.action_compare_to_example.setEnabled(False)
+
+        self.action_show_example.setCheckable(True)
+        self.action_show_example.setChecked(False)
+
+        ### Export menu actions
         self.action_export_analysis = QtGui.QAction(main_window)
         self.action_export_analysis.setEnabled(False)
         self.action_export_analysis.setObjectName("action_export_analysis")
@@ -236,15 +269,21 @@ class UiMainWindow(object):
             "action_export_distance_matrices"
         )
 
-        self.menu_file.addAction(self.action_open)
-        self.menu_file.addAction(self.action_save_current_textgrid)
-        self.menu_file.addAction(self.action_save_all_textgrids)
-        self.menu_file.addAction(self.action_save_all)
-        self.menu_file.addSeparator()
-        self.menu_file.addAction(self.action_quit)
+        self.menu_export.addAction(self.action_export_aggregate_images)
+        self.menu_export.addAction(self.action_export_annotations_and_metadata)
+        self.menu_export.addAction(self.action_export_distance_matrices)
+        self.menu_export.addAction(self.action_export_main_figure)
+        self.menu_export.addAction(self.action_export_ultrasound_frame)
 
-        # self.menu_script.addAction(self.actionShow_interpreter)
-        # self.menu_script.addAction(self.actionRun_file)
+        ### Navigation menu actions
+        self.actionNext = QtGui.QAction(main_window)
+        self.actionNext.setObjectName("actionNext")
+        self.actionPrevious = QtGui.QAction(main_window)
+        self.actionPrevious.setObjectName("actionPrevious")
+        self.actionNext_Frame = QtGui.QAction(main_window)
+        self.actionNext_Frame.setObjectName("actionNext_Frame")
+        self.actionPrevious_Frame = QtGui.QAction(main_window)
+        self.actionPrevious_Frame.setObjectName("actionPrevious_Frame")
 
         self.menu_navigation.addAction(self.actionNext)
         self.menu_navigation.addAction(self.actionPrevious)
@@ -252,17 +291,22 @@ class UiMainWindow(object):
         self.menu_navigation.addAction(self.actionNext_Frame)
         self.menu_navigation.addAction(self.actionPrevious_Frame)
 
-        self.menu_export.addAction(self.action_export_aggregate_images)
-        self.menu_export.addAction(self.action_export_annotations_and_metadata)
-        self.menu_export.addAction(self.action_export_distance_matrices)
-        self.menu_export.addAction(self.action_export_main_figure)
-        self.menu_export.addAction(self.action_export_ultrasound_frame)
+        ### Script menu actions
+        self.actionShow_interpreter = QtGui.QAction(main_window)
+        self.actionShow_interpreter.setObjectName("actionShow_interpreter")
+        self.actionRun_file = QtGui.QAction(main_window)
+        self.actionRun_file.setObjectName("actionRun_file")
 
+        # self.menu_script.addAction(self.actionShow_interpreter)
+        # self.menu_script.addAction(self.actionRun_file)
+
+        ### Menubar setup
         self.menubar.addAction(self.menu_file.menuAction())
+        self.menubar.addAction(self.menu_exercise.menuAction())
         self.menubar.addAction(self.menu_export.menuAction())
         self.menubar.addAction(self.menu_image.menuAction())
-        self.menubar.addAction(self.menu_script.menuAction())
         self.menubar.addAction(self.menu_navigation.menuAction())
+        self.menubar.addAction(self.menu_script.menuAction())
 
         self.retranslateUi(main_window)
         QtCore.QMetaObject.connectSlotsByName(main_window)
@@ -280,11 +324,29 @@ class UiMainWindow(object):
         self.positionRB_2.setText(_translate("MainWindow", "Low"))
         self.positionRB_3.setText(
             _translate("MainWindow", "Other / Not visible"))
+
         self.menu_file.setTitle(_translate("MainWindow", "File"))
-        self.menu_image.setTitle(_translate("MainWindow", "Image"))
-        self.menu_script.setTitle(_translate("MainWindow", "Script"))
-        self.menu_navigation.setTitle(_translate("MainWindow", "Navigation"))
+        self.menu_exercise.setTitle(_translate("MainWindow", "Exercise"))
         self.menu_export.setTitle(_translate("MainWindow", "Export"))
+        self.menu_image.setTitle(_translate("MainWindow", "Image"))
+        self.menu_navigation.setTitle(_translate("MainWindow", "Navigation"))
+        self.menu_script.setTitle(_translate("MainWindow", "Script"))
+
+        self.action_run_as_exercise.setText(
+            _translate("MainWindow", "Run as exercise"))
+        self.action_create_exercise.setText(
+            _translate("MainWindow", "Create exercise..."))
+        self.action_open_exercise.setText(
+            _translate("MainWindow", "Open exercise..."))
+        self.action_open_answer.setText(
+            _translate("MainWindow", "Open answer..."))
+        self.action_save_answer.setText(
+            _translate("MainWindow", "Save answer..."))
+        self.action_compare_to_example.setText(
+            _translate("MainWindow", "Compare to example"))
+        self.action_show_example.setText(
+            _translate("MainWindow", "Show example"))
+
         self.actionNew.setText(_translate("MainWindow", "New"))
         self.action_open.setText(_translate("MainWindow", "Open..."))
         self.action_open.setShortcut(_translate("MainWindow", "Ctrl+O"))
@@ -335,6 +397,14 @@ class UiMainWindow(object):
         )
 
     def add_items_to_database_view(self, session: Session):
+        """
+        Add items/recordings to the list view.
+
+        Parameters
+        ----------
+        session : Session
+            Use the recordings in the given session to populate the list view.
+        """
         for recording in session:
             self.database_model.appendRow(
                 QtGui.QStandardItem(
@@ -342,3 +412,15 @@ class UiMainWindow(object):
                     f"{recording.metadata.prompt.strip()}"
                 )
             )
+
+    def replace_items_in_database_view(self, session: Session):
+        """
+        Replace the items/recordings in the list view.
+
+        Parameters
+        ----------
+        session : Session
+            Use the recordings in the given Session to replace the old ones. 
+        """
+        self.database_model.clear()
+        self.add_items_to_database_view(session)