+++ src/patkit/qt_annotator.py
@@ -89,7 +89,9 @@ from patkit.plot_and_publish import (
     plot_wav,
 )
 from patkit.plot_and_publish.plot import plot_spectrogram2
-from patkit.save_and_load import load_recording_session, save_recording_session
+from patkit.save_and_load import (
+    load_exercise, load_recording_session, save_recording_session
+)
 from patkit.ui_callbacks import UiCallbacks
 
 _logger = logging.getLogger('patkit.qt_annotator')
@@ -130,19 +132,20 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
         super().__init__()
         self.kymography_clicker = None
         self.setupUi(self)
-
-        self.add_items_to_database_view(session)
         setup_qtannotator_ui_callbacks()
 
         self.session = session
         self.recordings = session.recordings
         self.index = 0
+
         self.max_index = len(self.recordings)
 
         self.display_tongue = display_tongue
 
         self.data_config = config.data_config
         self.gui_config = config.gui_config
+
+        self.add_items_to_database_view(session)
 
         if categories is None:
             self.categories = PdQtAnnotator.default_categories
@@ -225,6 +228,16 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
         self.action_save_current_textgrid.triggered.connect(self.save_textgrid)
         self.action_save_all_textgrids.triggered.connect(
             self.save_all_textgrids)
+
+        self.action_run_as_exercise.triggered.connect(
+            self.run_as_exercise)
+        self.action_create_exercise.triggered.connect(
+            self.create_exercise)
+        self.action_open_exercise.triggered.connect(self.open_exercise)
+        self.action_open_answer.triggered.connect(self.open_answer)
+        self.action_save_answer.triggered.connect(self.save_answer)
+        self.action_compare_to_example.triggered.connect(self.compare_to_model)
+        self.action_show_example.triggered.connect(self.show_model)
 
         self.action_export_aggregate_images.triggered.connect(
             self.export_aggregate_image)
@@ -362,7 +375,7 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
 
     @property
     def current(self):
-        """Current recording index."""
+        """Current recording at index."""
         return self.recordings[self.index]
 
     @property
@@ -373,6 +386,7 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
             'tonguePosition': self.tongue_positions[-1],
             'selected_time': -1.0,
             'selection_index': -1,
+            'frame_selection_index': -1,
         }
 
     def _add_annotations(self):
@@ -555,6 +569,11 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
                 self._get_long_title() + "\nNOTE: Audio missing.")
             return
 
+        # TODO 0.18.1: Add a check to draw plots which adds the model textgrid to
+        # plotting 
+        if self.action_show_example.isChecked():
+            print("I should be showing the model answer but don't yet know how.")
+
         for axes in self.tier_axes:
             axes.remove()
         self.tier_axes = []
@@ -629,10 +648,13 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
                         mode=self.gui_config.color_scheme
                     )
                 case "wav":
-                    plot_wav(ax=self.data_axes[axes_counter],
-                             waveform=wav,
-                             wav_time=wav_time,
-                             xlim=self.xlim)
+                    plot_wav(
+                        ax=self.data_axes[axes_counter],
+                        waveform=wav,
+                        wav_time=wav_time,
+                        xlim=self.xlim,
+                        mode=self.gui_config.color_scheme
+                    )
                 case _:
                     if not self.current.excluded:
                         self.plot_modality_axes(
@@ -709,8 +731,16 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
         Display an already interpolated ultrasound frame.
         """
         # Display mean image if asked or if there is no selection cursor.
-        if (self.current.annotations['selection_index'] == -1
-                or self.image_type == GuiImageType.MEAN_IMAGE):
+        if 'RawUltrasound' not in self.current.modalities:
+            return
+
+        if (
+            (
+                'frame_selection_index' not in self.current.annotations or
+                self.current.annotations['frame_selection_index'] == -1
+            )
+            or self.image_type == GuiImageType.MEAN_IMAGE
+        ):
             self.action_export_ultrasound_frame.setEnabled(False)
             self.ultra_axes.clear()
             image_name = 'AggregateImage mean on RawUltrasound'
@@ -722,10 +752,13 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
                     extent=(-image.shape[1] / 2 - .5, image.shape[1] / 2 + .5,
                             -.5, image.shape[0] + .5))
         # Display either raw or interpolated ultrasound if asked
-        elif self.current.annotations['selection_index'] >= 0:
+        elif (
+            'frame_selection_index' in self.current.annotations and
+            self.current.annotations['frame_selection_index'] >= 0
+        ):
             self.action_export_ultrasound_frame.setEnabled(True)
             self.ultra_axes.clear()
-            index = self.current.annotations['selection_index']
+            index = self.current.annotations['frame_selection_index']
 
             ultrasound = self.current.modalities['RawUltrasound']
             if self.image_type == GuiImageType.FRAME:
@@ -752,7 +785,7 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
             if (self.image_type == GuiImageType.FRAME
                     and 'Splines' in self.current.modalities):
                 splines = self.current.modalities['Splines']
-                index = self.current.annotations['selection_index']
+                index = self.current.annotations['frame_selection_index']
                 ultra = self.current.modalities['RawUltrasound']
                 timestamp = ultra.timevector[index]
 
@@ -802,9 +835,9 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
         """
         Interpolate and display a raw ultrasound frame.
         """
-        if self.current.annotations['selection_index'] > -1:
+        if self.current.annotations['frame_selection_index'] > -1:
             self.action_export_ultrasound_frame.setEnabled(True)
-            ind = self.current.annotations['selection_index']
+            ind = self.current.annotations['frame_selection_index']
             array = self.current.modalities['RawUltrasound'].data[ind, :, :]
         else:
             self.action_export_ultrasound_frame.setEnabled(False)
@@ -837,27 +870,29 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
         else:
             stimulus_onset = audio.go_signal
 
+        # TODO 0.19: This should not be hard coded
         if 'PD l1 on RawUltrasound' in self.current.modalities:
             pd_metrics = self.current.modalities['PD l1 on RawUltrasound']
             ultra_time = pd_metrics.timevector - stimulus_onset
-            index = self.current.annotations['selection_index']
+            index = self.current.annotations['frame_selection_index']
             self.current.annotations['selected_time'] = ultra_time[index]
 
     def next_frame(self):
         """
         Move the data cursor to the next frame.
         """
+        # TODO 0.19: Remove hard coding again
         if 'PD l1 on RawUltrasound' not in self.current.modalities:
             return
 
-        selection_index = self.current.annotations['selection_index']
+        frame_selection_index = self.current.annotations['frame_selection_index']
         pd = self.current.modalities['PD l1 on RawUltrasound']
         data_length = pd.data.size
-        if -1 < selection_index < data_length:
-            self.current.annotations['selection_index'] += 1
+        if -1 < frame_selection_index < data_length:
+            self.current.annotations['frame_selection_index'] += 1
             _logger.debug(
                 "next frame: %d",
-                (self.current.annotations['selection_index']))
+                (self.current.annotations['frame_selection_index']))
             self._update_pd_onset()
             self.update()
             self.update_ui()
@@ -866,11 +901,11 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
         """
         Move the data cursor to the previous frame.
         """
-        if self.current.annotations['selection_index'] > 0:
-            self.current.annotations['selection_index'] -= 1
+        if self.current.annotations['frame_selection_index'] > 0:
+            self.current.annotations['frame_selection_index'] -= 1
             _logger.debug(
                 "previous frame: %d",
-                (self.current.annotations['selection_index']))
+                (self.current.annotations['frame_selection_index']))
             self._update_pd_onset()
             self.update()
             self.update_ui()
@@ -938,13 +973,26 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
         directory = QFileDialog.getExistingDirectory(
             self, caption="Open directory", directory='.')
         if directory:
+            # TODO 0.18.1: these should be loaded from the new directory as well
+            # self.display_tongue = display_tongue
+
+            # self.data_config = config.data_config
+            # self.gui_config = config.gui_config
+
             self.session = load_recording_session(
                 directory=directory)
+            for recording in self.session:
+                recording.after_modalities_init()
             self.recordings = self.session.recordings
             self.index = 0
             self.max_index = len(self.recordings)
+            go_validator = QIntValidator(1, self.max_index + 1, self)
+            self.go_to_line_edit.setValidator(go_validator)
+            self.replace_items_in_database_view(session=self.session)
             self._add_annotations()
+
             self.update()
+            self.update_ui()
 
     def open_file(self):
         """
@@ -1020,6 +1068,47 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
                     "Wrote TextGrid to file %s.",
                     str(recording.textgrid_path))
 
+
+    def run_as_exercise(self):
+        """
+        Scramble TextGrids to run as an Exercise.
+        """
+        for recording in self.session:
+            for tier in recording.patgrid:
+                recording.patgrid[tier].scramble()
+        self.update()
+        self.update_ui()
+
+    def create_exercise(self):
+        """
+        Wrap a directory as an Exercise.
+        """
+        # TODO 2.6.: JOS MAHDOLLISTA ENNEN EMBRAN ESITELMÄÄ
+        # ask for directory
+        # ask for patkit/exercise dir
+        # write patkit_v.yaml in exercise dir
+        # mess up the textgrids as equidistant
+        # show scrambled textgrid instead of original
+
+    def open_exercise(self):
+        # (exercise_config_path, _) = QFileDialog.getOpenFileName(
+        #     self, 'Open Exercise', directory='.',
+        #     filter="Exercise files (patkit_exercise.yaml)")
+        # self.exercise = load_exercise(exercise_config_path)
+        pass
+
+    def open_answer(self):
+        pass
+
+    def save_answer(self):
+        pass
+
+    def compare_to_model(self):
+        print("Comparing to model has not yet been implemented.")       
+
+    def show_model(self):
+        self.update()
+
     def export_figure(self):
         """
         Callback method to export the current figure in any supported format.
@@ -1052,7 +1141,7 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
         # TODO: Add a check that grays out the export ultrasound figure when
         # one isn't available.
 
-        if self.current.annotations['selection_index'] >= 0:
+        if self.current.annotations['frame_selection_index'] >= 0:
             suggested_path = Path.cwd() / "Raw_ultrasound_frame.png"
             path, options = ImageSaveDialog.get_selection(
                 name="Export ultrasound frame",
@@ -1073,7 +1162,7 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
                 filepath=path,
                 session=self.session,
                 recording=self.current,
-                selection_index=self.current.annotations['selection_index'],
+                selection_index=self.current.annotations['frame_selection_index'],
                 selection_time=self.current.annotations['selected_time'],
                 ultrasound=self.current['RawUltrasound'],
                 interpolation_params=interpolation_params
@@ -1293,6 +1382,7 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
         if not event.xdata:
             self.current.annotations['selected_time'] = -1
             self.current.annotations['selection_index'] = -1
+            self.current.annotations['frame_selection_index'] = -1
             self.update()
             return
 
@@ -1315,15 +1405,24 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
         else:
             stimulus_onset = audio.go_signal
 
-        timevector = (
-            self.current.modalities['PD l1 on RawUltrasound'].timevector)
-        distances = np.abs(timevector - stimulus_onset - event.xdata)
-        self.current.annotations['selection_index'] = np.argmin(distances)
-        self.current.annotations['selected_time'] = event.xdata
+        # TODO 0.19: Remove hardcoding of modality names?
+        if 'RawUltrasound' in self.current.modalities:
+            timevector = (
+                self.current.modalities['RawUltrasound'].timevector)
+            distances = np.abs(timevector - stimulus_onset - event.xdata)
+            self.current.annotations['frame_selection_index'] = np.argmin(distances)
+            self.current.annotations['selected_time'] = event.xdata
+        if 'MonoAudio' in self.current.modalities:
+            timevector = (
+                self.current.modalities['MonoAudio'].timevector)
+            distances = np.abs(timevector - stimulus_onset - event.xdata)
+            self.current.annotations['selection_index'] = np.argmin(distances)
+            self.current.annotations['selected_time'] = event.xdata
 
         _logger.debug(
-            "Inside onpick - subplot: %d, index=%d, x=%f",
+            "Inside onpick - subplot: %d, ultra_index=%d, audio_index=%d, x=%f",
             subplot,
+            self.current.annotations['frame_selection_index'],
             self.current.annotations['selection_index'],
             self.current.annotations['selected_time'])
 
@@ -1402,6 +1501,8 @@ class PdQtAnnotator(QMainWindow, UiMainWindow):
         """
         self.gui_config.auto_xlim = False
         if self.current.annotations['selection_index'] >= 0:
+            center = self.current.annotations['selected_time']
+        elif self.current.annotations['frame_selection_index'] >= 0:
             center = self.current.annotations['selected_time']
         else:
             center = (self.xlim[0] + self.xlim[1]) / 2.0